% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_date_data}
\alias{make_date_data}
\title{Make tidy data}
\usage{
make_date_data(dir.data, fn.data, fn.record, dir.output, fn.output,
  .guess_max = 200)
}
\arguments{
\item{dir.data}{Character}

\item{fn.data}{Character}

\item{fn.record}{Charater}

\item{dir.output}{Character}

\item{fn.output}{Character}
}
\description{
Make tidy data
}
